+++ app/modules/Hello/routes.js
@@ -1,0 +1,11 @@
+/**
+ * Hello Router
+ */
+'use strict'
+
+const hello = require('./controller')
+const router = require('express').Router()
+
+router.get('/', hello.index)
+
+module.exports = router