+++ app/routes.js
@@ -1,0 +1,14 @@
+/**
+ * Router
+ */
+'use strict'
+
+const hello = require('./modules/Hello')
+
+module.exports = app => {
+  app.use('/hello', hello)
+
+  app.get('/', (request, response) => {
+    response.render('./views/welcome')
+  })
+}