+++ public/assets/scripts/main.js
@@ -1,0 +1,8 @@
+/**
+ * Main JS
+ */
+(function () {
+  'use strict'
+
+  console.log('Hello from express boilerplate')
+})(window, document)