+++ CHANGELOG.md
@@ -2,5 +2,8 @@
 
 ## Unreleased
 
-- Create hello módule
-- Added basics módules
+- Added public to static files
+- Added main route file
+- Added route file to module
+- Create hello module
+- Added basics modules